+++ .github/CHANGELOG.md
@@ -7,8 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Adjusted build file to reference PSScriptAnalyzerSettings file
+- Added PSScriptAnalyzerSettings.psd1 default Script Analyzer settings file
+- Added option for specifying PowerShell coding style
+- Added extensions.json file recommending PowerShell and Code Coverage Gutters extension
+- Removed option for .gitignore file. It is now included by default.
+- Added Code of Conduct file option
 - Added License file option
-- Updated all references to platyPS to explicitly use v.0.12.0 of platyPS
+- Updated all references to platyPS to explicitly use v0.12.0 of platyPS
 - Added Changelog file option
 - Added GitHub community files:
   - Code of Conduct